+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## lifebit-ai/cloudos-cli: changelog
 
+## v2.24.0 (2025-05-29)
+
+- Adds `--azure-worker-instance-type` to `cloudos job run` command, to be able to specify the worker node instance type to be used in azure.
+- Adds `--azure-worker-instance-disk` to `cloudos job run` command, to be able to specify the disk size in GB for the worker node to be used in azure.
+- Adds `--azure-worker-instance-spot` to `cloudos job run` command, to be able to specify whether the azure worker nodes have to be spot instances or not.
+
+
 ## v2.23.0 (2025-05-26)
 
 - Updates jobs POST endpoint from v1 to v2